import: rename import crate to sst_importer
